added untracked files, change file structure for heroku deploy
