+++ package.json
@@ -3,12 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "main": "server.js",
+  "babel": {
+  "presets": [
+    "es2015",
+    "stage-0"
+    ]
+  },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "apitest": "mocha",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "start": "concurrently \"npm run server\" \"npm run client\"",
+    "server": "node server.js",
+    "client": "babel-node start-client.js"
   },
   "devDependencies": {
     "mocha": "^3.2.0",
@@ -18,6 +22,10 @@
   },
   "dependencies": {
     "axios": "^0.15.3",
+    "babel-cli": "6.6.5",
+    "babel-core": "6.4.5",
+    "babel-preset-es2015": "6.3.13",
+    "babel-preset-stage-0": "6.5.0",
     "body-parser": "^1.16.0",
     "express": "^4.14.0",
     "lodash": "^4.17.2",
@@ -28,6 +36,5 @@
   },
   "engines": {
     "node": "7.0.0"
-  },
-  "proxy": "http://localhost:3001/api"
+  }
 }
