--- public/index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="icon" type="image/x-icon" href="./code.png"/>
-    <!-- code icon by ProSymbols from the Noun Project -->
-    <link href="https://fonts.googleapis.com/css?family=Roboto:400,700" rel="stylesheet">
-    <title>Biggie/Smalls</title>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
