+++ server.js
@@ -9,7 +9,12 @@ const regexp = /(ftp|http|https):\/\/(\w+:{0,1}\w*@)?(\S+)(:[0-9]+)?(\/|\/([\w#!
 app.set('port', process.env.PORT || 3001);
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended:true }));
-app.use(express.static(__dirname + '/public'));
+
+// app.use(express.static(__dirname + '/public'));
+if (process.env.NODE_ENV === 'production') {
+  app.use(express.static('client/build'));
+}
+
 app.locals.urls = [];
 
 app.get('/api/urls', (request, response) => {
