--- src/css/App.css
@@ -1,114 +1,0 @@
-@font-face {
-  font-family: 'freshPrince';
-  src: url('./the-fresh-prince.otf');
-}
-
-.App {
-  font-family: 'Roboto', sans-serif;
-  text-align: center;
-}
-
-.title {
-  font-family: 'freshPrince', sans-serif;
-  font-size: 8em;
-  letter-spacing: 7px;
-  margin-top: 10px;
-  margin-bottom: 30px;
-  text-shadow: 1px 1px 1px white,
-               3px 3px 5px #FF5722;
-}
-
-.input-field {
-  border: 2px solid #FF5722;
-  height: 20px;
-  width: 300px;
-  margin-right: 10px;
-  text-indent: 5px;
-}
-
-.submit-button {
-  background: #FF5722;
-  border: 2px solid #FF5722;
-  border-radius: 15px;
-  color: white;
-  padding: 6px;
-}
-
-.submit-button:hover {
-  background: white;
-  border: 2px solid #FF5722;
-  color: #FF5722;
-}
-
-a {
-  color: black;
-  text-decoration: none;
-}
-
-a:hover {
-  color: #FF5722;
-}
-
-.short-url {
-  margin-bottom: 30px;
-  margin-top: 30px;
-}
-
-.list-container {
-  margin-top: 30px;
-}
-
-.search-container {
-  align-items: center;
-  display: inline-flex;
-}
-
-.search-field {
-  border: 2px solid #FF5722;
-  height: 20px;
-  width: 200px;
-  text-indent: 5px;
-}
-
-.pop-sort {
-  display: inline-flex;
-  flex-direction: row;
-}
-
-.date-sort {
-  align-items: center;
-  display: inline-flex;
-  flex-direction: column;
-}
-
-.sort-button {
-  background: #FF5722;
-  border-bottom: 1px solid #FF5722;
-  border-radius: 15px;
-  color: white;
-  margin-bottom: 10px;
-  margin-left: 10px;
-  width: 100px;
-}
-
-.sort-button:hover {
-  background: white;
-  border: 2px solid #FF5722;
-  color: #FF5722;
-}
-
-.url-table {
-  margin: auto;
-  margin-top: 20px;
-}
-
-.table-headers {
-  border-bottom: 1px solid #FF5722;
-  margin-left: 20px;
-  padding-bottom: 5px;
-  width: 150px;
-}
-
-.url-row {
-  height: 35px;
-}
