+++ start-client.js
@@ -1,0 +1,3 @@
+const args = [ 'start' ];
+const opts = { stdio: 'inherit', cwd: 'client', shell: true };
+require('child_process').spawn('npm', args, opts);
